pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transcriptor"
version = "0.1.0"
description = "CLI de transcription OCR pour images et PDF vers PDF texte"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "reportlab>=4.0.0",
    "pdf2image>=1.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
transcriptor = "transcriptor.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
